+++ src/Writer/ClusterWriter.php
@@ -54,9 +54,9 @@ class ClusterWriter implements WriterInterface
     public function save(
         ObjectInterface $object,
         PromiseInterface $promise = null,
-        ?bool $prefereRealDateOnUpdate = null,
+        ?bool $preferRealDateOnUpdate = null,
     ): WriterInterface {
-        $this->persist($object, $promise, $prefereRealDateOnUpdate);
+        $this->persist($object, $promise, $preferRealDateOnUpdate);
 
         return $this;
     }